Updates for PR review
