+++ Wire-iOS/Sources/UserInterface/Conversation/Content/Cells/LocationMessageCell.swift
@@ -212,35 +212,3 @@ public final class LocationMessageCell: ConversationCell {
     }
 }
 
-public extension ZMLocationMessageData {
-
-    func openInMaps(with span: MKCoordinateSpan) {
-        let launchOptions = [
-            MKLaunchOptionsMapCenterKey: NSValue(mkCoordinate: coordinate),
-            MKLaunchOptionsMapSpanKey: NSValue(mkCoordinateSpan: span)
-        ]
-
-        if let url = googleMapsURL, url.openAsLocation() {
-            return
-        }
-
-        mapItem?.openInMaps(launchOptions: launchOptions)
-    }
-
-    var googleMapsURL: URL? {
-        let location = "\(coordinate.latitude),\(coordinate.longitude)"
-        return URL(string: "comgooglemaps://?q=\(location)&center=\(location)&zoom=\(zoomLevel)")
-    }
-
-    var mapItem: MKMapItem? {
-        var addressDictionary: [String : AnyObject]? = nil
-        if let name = name {
-            addressDictionary = [String(kABPersonAddressStreetKey): name as AnyObject]
-        }
-        
-        let placemark = MKPlacemark(coordinate: coordinate, addressDictionary: addressDictionary)
-        let mapItem = MKMapItem(placemark: placemark)
-        mapItem.name = name
-        return mapItem
-    }
-}
